:sparkles: feat: removendo cadastroAdmin e cadastroUsuario e criando pagina Cadastro
