--- src/pages/cadastroAdmin/CadastroAdmin.tsx
@@ -1,11 +1,0 @@
-
-import Cadastro from "../components/cadastro/Cadastro";
-
-
-
-const CadastroAdmin = () => {
-   return <Cadastro admin={true} mensagemH1="Cadastro de Administrador"/>
-}
-
-export default CadastroAdmin;
-
