--- src/pages/cadastroUsuario/CadastroUsuario.tsx
@@ -1,11 +1,0 @@
-
-import Cadastro from "../components/cadastro/Cadastro";
-
-
-
-const CadastroUsuario = () => {
-   return <Cadastro admin={false} mensagemH1="Cadastro de Usuário"/>
-}
-
-export default CadastroUsuario;
-
