+++ src/pages/explorarMinhasPautas/ExplorarMinhasPautas.tsx
@@ -1,11 +1,17 @@
 import { useBuscarTodasPautasUsuarioLogado } from "../../hooks/useBuscarTodasPautasUsuarioLogado";
+import { useDadosUsuarioStore } from "../../hooks/useDadosUsuarioStore";
 import ExplorarPautas from "../components/explorarPautas/ExplorarPautas"
+import NaoAutorizado from "../components/naoAutorizado/NaoAutorizado";
 
 const ExplorarMinhasPautas = ()=>{
     const {buscarTodasPautasUsuarioLogado} = useBuscarTodasPautasUsuarioLogado()
+    const { admin } = useDadosUsuarioStore();
     return (
-        <ExplorarPautas metodoBuscarPautasBanco={buscarTodasPautasUsuarioLogado}/>
+        admin ? <ExplorarPautas metodoBuscarPautasBanco={buscarTodasPautasUsuarioLogado}/> : <NaoAutorizado/>
     )
+        
+        
+    
 }
 
 export default ExplorarMinhasPautas;
