+++ src/routes/Routes.tsx
@@ -1,32 +1,28 @@
 import { RouterProvider, createBrowserRouter } from "react-router-dom";
-import CadastroUsuario from "../pages/cadastroUsuario/CadastroUsuario";
 import Login from "../pages/login/Login";
-import CadastroAdmin from "../pages/cadastroAdmin/CadastroAdmin";
 import RotasComCabecalho from "./RotasComCabecalho";
 import RotasAutenticadas from "./RotasAutenticadas";
 import ExplorarPautasAtivas from "../pages/explorarPautasAtivas/ExplorarPautasAtivas";
 import ExplorarMinhasPautas from "../pages/explorarMinhasPautas/ExplorarMinhasPautas";
+import Cadastro from "../pages/cadastro/Cadastro";
 
 const router = createBrowserRouter([
     {
         path: "/",
         element: <Login />,
     },
+    
     {           
         element: <RotasAutenticadas />,
         children: [
-            {
-                path: "/novoAdmin",
-                element: <CadastroAdmin />,
-            },
-            {
-                path: "/novoUsuario",
-                element: <CadastroUsuario />,
-            },
             
             {
                 element: <RotasComCabecalho />,
                 children: [
+                    {
+                        path: "/cadastro",
+                        element: <Cadastro />,
+                    },
                     {
                         path: "/explorar",
                         element: <ExplorarPautasAtivas />,
